Creado producto y etapa

+++ avicola_backend/src/main/java/com/wil/avicola_backend/service/StageService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import com.wil.avicola_backend.error.RequestException;
 import com.wil.avicola_backend.model.Stage;
+import com.wil.avicola_backend.repository.ProductRepository;
 import com.wil.avicola_backend.repository.StageRepository;
 
 @Service
@@ -14,14 +15,17 @@ public class StageService {
 
     @Autowired
     private StageRepository stageRepository;
+    @Autowired
+    private ProductRepository productRepository;
 
     public ResponseEntity<?> findStages() {
         return ResponseEntity.ok().body(stageRepository.findAll());
     }
 
     public ResponseEntity<Stage> saveStage(Stage stage) {
+
         if (stageRepository.existsByName(stage.getName())) {
-            throw new RequestException("Ya existe una etapa con el mismo nombre.");
+            throw new RequestException("Ya existe etapa con el mismo nombre.");
         }
 
         Stage stage_new = stageRepository.save(stage);
@@ -29,9 +33,10 @@ public class StageService {
     }
 
     public ResponseEntity<Stage> updateStage(Stage stage) {
+
         if (stageRepository.existsById(stage.getId())) {
             if (stageRepository.existsByNameOther(stage.getId(), stage.getName())) {
-                throw new RequestException("Ya existe una etapa con el mismo nombre.");
+                throw new RequestException("Ya existe etapa con el mismo nombre.");
             }
 
             Stage stage_old = stageRepository.findById(stage.getId()).get();
@@ -40,16 +45,22 @@ public class StageService {
             stageRepository.save(stage_old);
             return ResponseEntity.status(HttpStatus.OK).body(stage_old);
         }
-        throw new RequestException("No existe la etapa.");
+        throw new RequestException("No existe etapa.");
     }
 
     public ResponseEntity<Stage> deleteStage(long id) {
+
+        // Si esta relacionado con producto
+        if (productRepository.existsByStage(id)) {
+            throw new RequestException("No puede ser eliminado, existe product(s) registrados con la etapa.");
+        }
+
         if (stageRepository.existsById(id)) {
             Stage stage = stageRepository.findById(id).get();
             stageRepository.deleteById(id);
             return ResponseEntity.status(HttpStatus.OK).body(stage);
         }
 
-        throw new RequestException("No existe la etapa.");
+        throw new RequestException("No existe etapa.");
     }
 }
